Initial commit of exp_viz library
This is a copy of exp_viz from the cegs-portal repository. This
will get de-py03ed, and exp_viz in the cegs-portal repo will import
this and add all the necessary p03 stuff.

This is getting factored out now because cov_viz_manifest needs to
to some filtering to get the right initial values for the effect size
and significance sliders.